fix de merde

+++ vm/sources/reg_manager.c
@@ -30,7 +30,7 @@ int				num_at_reg(t_vm vm, int pos, size_t size)
 	return (n);
 }
 
-void			set_num_at_reg(t_vm *vm, t_pro pro, int reg_pos, size_t reg)
+void			set_num_at_reg(t_vm *vm, t_pro pro, int reg_pos, unsigned int reg)
 {
 	int		i;
